# docs/assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IBEF Backend API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/test/info": {
      "post": {
        "tags": [
          "test"
        ],
        "summary": "Set Test Info",
        "description": "Set test metadata to prepare for a test run.\n\nThis endpoint stores the test information without starting data collection.\nAfter calling this, the test will be in PREPARED state.\nThen call PUT /start to begin the actual test.",
        "operationId": "set_test_info_api_test_info_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestMetaData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "409": {
            "description": "A test is already running. Cannot prepare new metadata while a test is in progress.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "A test is already running."
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/test/start": {
      "put": {
        "tags": [
          "test"
        ],
        "summary": "Start Test",
        "description": "Start the prepared test and begin recording data.\n\nRequires that metadata has been set via POST /info first.\nChanges state from PREPARED to RUNNING.",
        "operationId": "start_test_api_test_start_put",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "No test metadata prepared. Call POST /info first.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "No test metadata prepared. Call POST /info first."
                }
              }
            }
          },
          "409": {
            "description": "A test is already running.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "A test is already running."
                }
              }
            }
          }
        }
      }
    },
...
    "/api/test/description": {
      "get": {
        "tags": [
          "test"
        ],
        "summary": "Get Current Test Description",
        "description": "Get the description of the current test (PREPARED or RUNNING state).\n\nAvailable after POST /info has been called.\nReturns the markdown content of description.md",
        "operationId": "get_current_test_description_api_test_description_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Current Test Description Api Test Description Get"
                }
              }
            }
          },
          "409": {
            "description": "No test prepared. Call POST /info first.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "No test prepared. Call POST /info first."
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "test"
        ],
        "summary": "Update Current Test Description",
        "description": "Update the description of the current test (PREPARED or RUNNING state).\n    \n    Available after POST /info has been called.\n    Modifies the description.md file in real-time.\n    \n    Request body:\n    ```json\n    {\n        \"content\": \"# Updated Description\n\nNew markdown content...\"\n    }\n    ```",
        "operationId": "update_current_test_description_api_test_description_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "409": {
            "description": "No test prepared. Call POST /info first.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "No test prepared. Call POST /info first."
                }
              }
            }
          },
          "500": {
            "description": "Failed to update description.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Failed to update description: Permission denied"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TestMetaData": {
        "properties": {
          "test_id": {
            "type": "string",
            "title": "Test Id"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "operator_name": {
            "type": "string",
            "title": "Operator Name"
          },
          "specimen_code": {
            "type": "string",
            "title": "Specimen Code"
          },
          "dim_length": {
            "type": "number",
            "title": "Dim Length",
            "default": 0.0
          },
          "dim_height": {
            "type": "number",
            "title": "Dim Height",
            "default": 0.0
          },
          "dim_width": {
            "type": "number",
            "title": "Dim Width",
            "default": 0.0
          },
          "loading_mode": {
            "type": "string",
            "title": "Loading Mode",
            "default": ""
          },
          "sensor_spacing": {
            "type": "number",
            "title": "Sensor Spacing",
            "default": 0.0
          },
          "ext_support_spacing": {
            "type": "number",
            "title": "Ext Support Spacing",
            "default": 0.0
          },
          "load_point_spacing": {
            "type": "number",
            "title": "Load Point Spacing",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "test_id",
          "date",
          "operator_name",
          "specimen_code"
        ],
        "title": "TestMetaData",
        "description": "Data class representing metadata for a test.\nCompatible with both dataclass operations and Pydantic validation."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}